move structs to structs
