+++ src/model/structures/mod.rs
@@ -1,6 +1,6 @@
 pub mod match_cost;
 pub mod mode;
 pub mod player_rating;
-pub mod rating_calculation_result;
+pub mod processing;
 pub mod scoring_type;
 pub mod team_type;
