--- src/model/structures/rating_calculation_result.rs
@@ -1,11 +1,0 @@
-use crate::{
-    api::api_structs::{MatchRatingStats, RatingAdjustment},
-    model::structures::player_rating::PlayerRating
-};
-
-#[derive(Debug)]
-pub struct RatingCalculationResult {
-    pub base_ratings: Vec<PlayerRating>,
-    pub rating_stats: Vec<MatchRatingStats>,
-    pub adjustments: Vec<RatingAdjustment>
-}
